# Conflicts:
#	src/main/kotlin/ar/edu/unsam/phm/magicnightsback/domain/Show.kt

+++ src/main/kotlin/ar/edu/unsam/phm/magicnightsback/domain/Location.kt
@@ -1,6 +1,6 @@
 package ar.edu.unsam.phm.magicnightsback.domain
 
-data class Location (
-    val latitud: Float,
-    val longitud: Float
+data class Location(
+    val latitud: Double,
+    val longitud: Double
 )
